All examples working correctly except for movement in FlippyFlop example.
