+++ .vscode/launch.json
@@ -10,7 +10,13 @@
       "type": "coreclr",
       "request": "launch",
       "preLaunchTask": "build",
-      "program": "${workspaceFolder}/Examples/AutoTilingExample/bin/Debug/netcoreapp3.0/AutoTilingExample.exe",
+      /*"program": "${workspaceFolder}/Examples/AutoTilingExample/bin/Debug/netcoreapp3.0/AutoTilingExample.exe",*/
+      /*"program": "${workspaceFolder}/Examples/Coroutine/bin/Debug/netcoreapp3.0/Couroutine.exe",*/
+      /*"program": "${workspaceFolder}/Examples/EditTextBox/bin/Debug/netcoreapp3.0/EditTextBox.exe",*/
+      /*"program": "${workspaceFolder}/Examples/FlippyFlop/bin/Debug/netcoreapp3.0/FlippyFlop.exe",*/
+      /*"program": "${workspaceFolder}/Examples/GridCollider/bin/Debug/netcoreapp3.0/GridCollider.exe",*/
+      "program": "${workspaceFolder}/Examples/OtterPongGame/bin/Debug/netcoreapp3.0/OtterPongGame.exe",
+      /*"program": "${workspaceFolder}/Examples/PlatformingMovement/bin/Debug/netcoreapp3.0/PlatformingMovement.exe",*/
       "args": [],
       "cwd": "${workspaceFolder}",
       "console": "internalConsole",
