+++ Examples/AutoTilingExample/Program.cs
@@ -1,9 +1,4 @@
 ﻿using Otter;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace AutoTilingExample {
     class Program {
