+++ Examples/PlatformingMovement/PlatformingMovement.csproj
@@ -4,6 +4,11 @@
     <ProjectReference Include="..\..\Otter\Otter.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <None Update="OgmoProject.oep" CopyToOutputDirectory="PreserveNewest" />
+    <None Update="Level.oel" CopyToOutputDirectory="PreserveNewest" />
+  </ItemGroup>
+
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.0</TargetFramework>
